+++ CHANGELOG.rst
@@ -4,7 +4,15 @@ Release notes
 V0.2.0 (In development)
 +++++++++++++++++++++++
 
-No changes yet.
+New features:
+
+* Added NORM_WEIGHTED flag for weighted normalization. Added :func:`.core.norm_flags`.
+* Updated :class:`.viewer.CorrViewer` to implement new normalization type.
+* Updated :func:`.core.normalize` to implement new normalization type.
+
+Bug fixes:
+
+* Fixes a bug in calculation with binning = 2 (random selection). 
 
 V0.1.2 (Apr 21 2020)
 ++++++++++++++++++++